# web/api-docs/swagger.yaml
basePath: /
definitions:
...
  github_com_cuihe500_vaulthub_internal_database_models.SafeUser:
    properties:
      created_at:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      role:
        type: string
      status:
        $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.UserStatus'
      updated_at:
        type: string
      username:
        type: string
      uuid:
        type: string
    type: object
...
  github_com_cuihe500_vaulthub_internal_database_models.UserStatus:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      UserStatusActive: 活跃
      UserStatusDisabled: 已禁用
      UserStatusLocked: 已锁定
    x-enum-descriptions:
    - 活跃
    - 已禁用
    - 已锁定
    x-enum-varnames:
    - UserStatusActive
    - UserStatusDisabled
    - UserStatusLocked
...
  github_com_cuihe500_vaulthub_internal_service.ListUsersResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
      users:
        items:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUser'
        type: array
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.UpdateUserRoleRequest:
    properties:
      role:
        enum:
        - admin
        - user
        - readonly
        type: string
    required:
    - role
    type: object
  github_com_cuihe500_vaulthub_internal_service.UpdateUserStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.UserStatus'
        maximum: 3
        minimum: 1
    required:
    - status
    type: object
...
  github_com_cuihe500_vaulthub_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      requestId:
        type: string
      timestamp:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@thankseveryone.top
    name: API Support
  description: |-
    VaultHub 是一个密钥管理系统，旨在安全地存储、管理和轮换加密密钥、API 密钥及其他敏感凭证。

    ## 特性
    - 安全的密钥存储和管理
    - 密钥轮换支持
    - 细粒度的访问控制
    - 审计日志

    ## 认证
    大部分接口需要JWT Bearer Token认证。
    1. 调用 /api/v1/auth/register 注册账号
    2. 调用 /api/v1/auth/login 获取 token
    3. 在请求头中添加 Authorization: Bearer {token}

    ## 响应格式
    所有接口返回统一的响应格式，HTTP状态码均为200，错误码在响应体的code字段中：
    ```json
    {
    "code": 0,
    "message": "success",
    "data": {},
    "requestId": "xxx",
    "timestamp": 1762269490888
    }
    ```
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/cuihe500/vaulthub
  title: VaultHub API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: 获取用户列表（需要管理员权限）。不传分页参数时全量导出（最多10000条）
      parameters:
      - description: 页码（可选，不传则全量导出）
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页数量（可选，不传则全量导出）
        in: query
        maximum: 10000
        minimum: 1
        name: page_size
        type: integer
      - description: 用户状态
        enum:
        - 1
        - 2
        - 3
        in: query
        name: status
        type: integer
      - description: 用户角色
        enum:
        - admin
        - user
        - readonly
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.ListUsersResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
  /api/v1/users/{uuid}:
    get:
      consumes:
      - application/json
      description: 根据UUID获取用户详细信息（需要管理员权限）
      parameters:
      - description: 用户UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUser'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 用户管理
  /api/v1/users/{uuid}/role:
    put:
      consumes:
      - application/json
      description: 更新用户角色（需要管理员权限）
      parameters:
      - description: 用户UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: 更新角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUser'
              type: object
      security:
      - BearerAuth: []
      summary: 更新用户角色
      tags:
      - 用户管理
  /api/v1/users/{uuid}/status:
    put:
      consumes:
      - application/json
      description: 更新用户状态（需要管理员权限）
      parameters:
      - description: 用户UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: 更新状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUser'
              type: object
      security:
      - BearerAuth: []
      summary: 更新用户状态
      tags:
      - 用户管理
...
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer" 后跟一个空格和 JWT token。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"